requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bleach==5.0.1
cdflib==0.4.7
certifi==2022.9.24
cffi==1.15.1
cftime==1.6.2
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
contourpy==1.0.5
cryptography==38.0.3
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Django==2.1.4
django-seed==0.3.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.1.0
Faker==15.0.0
fastapi==0.85.1
fastjsonschema==2.16.2
fonttools==4.37.4
greenlet==1.1.3.post0
h11==0.14.0
h5py==3.7.0
httptools==0.5.0
idna==3.4
iniconfig==1.1.1
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.10
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.19.1
jupyter_client==7.3.5
jupyterlab==3.4.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.15.2
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
marshmallow==3.18.0
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.4
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.1
nest-asyncio==1.5.6
netCDF4==1.6.1
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.3
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
Pillow==9.2.0
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg2==2.9.4
psycopg2-binary==2.9.4
psycopg2-pool==1.1
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.2.1
pywin32==304
pywinpty==2.0.8
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.0
requests==2.28.1
rsa==4.9
scipy==1.9.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spacepy==0.4.1
SQLAlchemy==1.4.42
stack-data==0.5.1
starlette==0.20.4
terminado==0.16.0
tinycss2==1.1.1
tomli==2.0.1
toposort==1.7
tornado==6.2
traitlets==5.4.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.17.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
websockets==10.3
widgetsnbextension==4.0.3
wincertstore==0.2
xarray==2022.9.0
